move translation logic into simpleI18n class
