+++ packages/i18n/src/simpleI18nUtils/tests/functionHelper.test.ts
@@ -1,33 +1,7 @@
 import {
   memoizedNumberFormatter,
   numberFormatCacheKey,
-  translate,
-  INTERPOLATE_FORMAT,
-} from '../translate';
-
-describe('INTERPOLATE_FORMAT', () => {
-  it.each([
-    // Matches
-    ['{foo}', ['foo']],
-    ['{f}', ['f']],
-    ['{{foo}}', ['foo']],
-    ['test {foo }', ['foo']],
-    ['{ foo} test', ['foo']],
-    ['test {     foo   } test', ['foo']],
-    ['test {foo} test {bar}', ['foo', 'bar']],
-    // No matches
-    ['{}', []],
-    ['{     }', []],
-    ['{ test {}', []],
-  ])('matches "%s" with keys %p', (input, keys) => {
-    expect.assertions(keys.length);
-    Array.from(input.matchAll(INTERPOLATE_FORMAT)).forEach(
-      ([_, key], matchIndex) => {
-        expect(key).toBe(keys[matchIndex]);
-      },
-    );
-  });
-});
+} from '../functionHelpers';
 
 describe('memoizedNumberFormatter()', () => {
   it.each`
@@ -184,95 +158,6 @@ describe('numberFormatCacheKey()', () => {
       `${locale}-${otherLocale}-{}`,
     );
   });
-});
-
-describe('translate()', () => {
-  const id = 'test';
-  const translations = {
-    // Lots of spaces to make sure we're taking the whole match length into account.
-    [id]: 'foo {    bar     } baz',
-    pluralizations: {
-      '0': 'I have no car!', // eslint-disable-line @typescript-eslint/naming-convention
-      '1': 'I have a single car!', // eslint-disable-line @typescript-eslint/naming-convention
-      one: 'I have {count} car!',
-      other: 'I have {count} cars!',
-      ordinal: {
-        one: 'This is my {ordinal}st car!',
-        two: 'This is my {ordinal}nd car!',
-        few: 'This is my {ordinal}rd car!',
-        other: 'This is my {ordinal}th car!',
-      },
-    },
-  };
-  const locale = 'en-CA';
-
-  it('returns selects the proper values for cardinal pluralization', () => {
-    const translationNone = translate(
-      'pluralizations',
-      {count: 0},
-      translations,
-      locale,
-    );
-
-    expect(translationNone).toBe('I have no car!');
-
-    const translationOne = translate(
-      'pluralizations',
-      {count: 1},
-      translations,
-      locale,
-    );
-    expect(translationOne).toBe('I have a single car!');
-
-    const translationOther = translate(
-      'pluralizations',
-      {count: 2},
-      translations,
-      locale,
-    );
-    expect(translationOther).toBe('I have 2 cars!');
-  });
-
-  it('returns selects the proper values for ordinal pluralization', () => {
-    const translationFirst = translate(
-      'pluralizations',
-      {ordinal: 1},
-      translations,
-      locale,
-    );
-    const translationSecond = translate(
-      'pluralizations',
-      {ordinal: 2},
-      translations,
-      locale,
-    );
-    const translationThird = translate(
-      'pluralizations',
-      {ordinal: 3},
-      translations,
-      locale,
-    );
-    const translationFourth = translate(
-      'pluralizations',
-      {ordinal: 4},
-      translations,
-      locale,
-    );
-    expect(translationFirst).toBe('This is my 1st car!');
-    expect(translationSecond).toBe('This is my 2nd car!');
-    expect(translationThird).toBe('This is my 3rd car!');
-    expect(translationFourth).toBe('This is my 4th car!');
-  });
-
-  it.each([2, 'bar'])(
-    'returns a string when a simple replacement is used',
-    (replacementValue) => {
-      const replacements = {bar: replacementValue};
-      const translation = translate('test', replacements, translations, locale);
-
-      expect(translation).toBe(`foo ${replacementValue} baz`);
-    },
-  );
 });
 
 function sanitizeSpaces(input) {
